In the design of wifidog, it implememt a polling mechanism of auth servers. It supports a chain of auth server, when it detects an auth server is unreachable, such as it could not resolve the host name of that server, it will shift the chain, putting the unreachable auth server at the end of the chain.
But this introduces a chain of complex mutex protection used since it is the only
condition that the auth server chain got modified.

We don't support mulitple auth server right now. So, removing this part.
But please be awared that you should be very very very careful if you want to add
it back.