In wifidog, in every cases that it makes desicions about a client, it should query auth server for confirmation. Such as, got token from the message redirected from client. checking client logout status periodly in clean_up thread...
Since we didn't use this mechanism, I just commented out function authenticate_client().

For logout_client(), I just need to delete the relative entry in iptables and clean up data
structure.

For the thread of client_timeout_check, the old way is sending out a icmp package to client
periodly, then check the statistic data from iptables counter log. It is not good if client
doesn't reply you.

In our implement, for each client got authed from server, server will give me a remain time
to record the alive time for him. So, I just need to count the time.
The design here is that this thread will run periodly to check the expired time for the clients
in the list. Setting up the next timer as the min(the closet expired time of clients, check interval).